feat: add jsonrepair util for Gemini responses
